Update `nodejs-npm-install` to use new layer API
