--- buildpacks/nodejs-npm-install/src/layers/mod.rs
@@ -1,2 +1,0 @@
-pub(crate) mod npm_cache;
-pub(crate) mod npm_runtime_config;
